Move WOLAPI and WOLBrowser libraries into Commando
No point to use such tiny code as separate library, moved into Commando code compilation.
