+++ Code/CMakeLists.txt
@@ -5,8 +5,6 @@ add_subdirectory(Commando)
 add_subdirectory(SControl)
 add_subdirectory(Scripts)
 add_subdirectory(Tools)
-add_subdirectory(wolapi)
-add_subdirectory(WOLBrowser)
 add_subdirectory(ww3d2)
 add_subdirectory(wwaudio)
 add_subdirectory(wwbitpack)
