+++ Code/WOLAPI/chatdefs.h
