+++ Code/WOLAPI/netutildefs.h
