+++ Code/WOLAPI/register.bat
