--- Code/WOLBrowser/CMakeLists.txt
@@ -1,10 +1,0 @@
-set(SOURCES
-    WOLBrowser_i.c
-)
-
-add_library(WOLBrowser STATIC ${SOURCES})
-target_include_directories(WOLBrowser PUBLIC
-    $<BUILD_INTERFACE:
-        ${PROJECT_SOURCE_DIR}/Code/WOLBrowser
-    >
-)
