+++ Code/WWOnline/WOLLoginInfo.h
@@ -39,10 +39,10 @@
 
 #include "RefCounted.h"
 #include "RefPtr.h"
-#include <WWLib\WideString.h>
+#include <WWLib/WideString.h>
 
 namespace WOL {
-#include <wolapi\wolapi.h>
+#include <WOLAPI/wolapi.h>
 }
 
 #if defined(_MSC_VER)
