--- Code/wolapi/CMakeLists.txt
@@ -1,10 +1,0 @@
-set(SOURCES
-    WOLAPI_i.c
-)
-
-add_library(WOLAPI STATIC ${SOURCES})
-target_include_directories(WOLAPI PUBLIC
-    $<BUILD_INTERFACE:
-        ${PROJECT_SOURCE_DIR}/Code/wolapi
-    >
-)
